requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
pycodestyle==2.10.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
